requirements.txt
amqp==5.1.1
anyio==3.6.2
asyncio==3.4.3
bidict==0.22.1
billiard==3.6.4.0
blobfile==2.0.1
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.3.0
fastapi==0.92.0
filelock==3.9.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.12.1
idna==3.4
joblib==1.2.0
kombu==5.2.4
lxml==4.9.2
nltk==3.8.1
numpy==1.24.2
OpenAIAuth==0.2.0
packaging==23.0
prompt-toolkit==3.0.36
pycryptodomex==3.17
pydantic==1.10.5
pymongo==4.3.3
python-dotenv==0.21.1
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
revChatGPT==2.3.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
starlette==0.25.0
tiktoken==0.2.0
tls-client==0.1.8
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
utils==1.0.1
uvicorn==0.20.0
vine==5.0.0
wa-automate-socket-client==1.2.0
wcwidth==0.2.6
websocket-client==1.5.1
